feat: add `testCppSolution` function from the main command
